+++ backend/src/routes/contactListRoutes.ts
@@ -16,6 +16,7 @@ routes.get("/contact-lists/:id", isAuth, ContactListController.show);
 routes.post("/contact-lists", isAuth, ContactListController.store);
 routes.post("/contact-lists/:id/upload",isAuth,upload.array("file"),ContactListController.upload);
 routes.post("/contact-lists/:id/import-system-contacts", isAuth, ContactListController.importSystemContacts);
+routes.post("/contact-lists/:id/revalidate-numbers", isAuth, ContactListController.revalidateNumbers);
 routes.put("/contact-lists/:id", isAuth, ContactListController.update);
 routes.delete("/contact-lists/:id", isAuth, ContactListController.remove);